refactor: remove `ref` postfix
