+++ apps/calendar/src/calendarContainer.tsx
@@ -2,7 +2,7 @@ import { FunctionComponent, h } from 'preact';
 
 import { StoreProvider } from '@src/contexts/calendarStore';
 import { EventBusProvider } from '@src/contexts/eventBus';
-import { FloatingLayerContainerProvider } from '@src/contexts/floatingLayerRef';
+import { FloatingLayerContainerProvider } from '@src/contexts/floatingLayer';
 import { ThemeProvider } from '@src/contexts/theme';
 import Theme from '@src/theme';
 import { EventBus } from '@src/utils/eventBus';
