+++ apps/calendar/src/components/dayGridMonth/gridCell.tsx
@@ -11,7 +11,7 @@ import {
   MONTH_MORE_VIEW_PADDING,
 } from '@src/constants/style';
 import { useDispatch } from '@src/contexts/calendarStore';
-import { useLayoutContainer } from '@src/contexts/layoutContainerRef';
+import { useLayoutContainer } from '@src/contexts/layoutContainer';
 import { useTheme } from '@src/contexts/theme';
 import { cls, toPercent } from '@src/helpers/css';
 import { getExceedCount } from '@src/helpers/grid';
