+++ apps/calendar/src/components/layout.tsx
@@ -6,7 +6,7 @@ import { EventDetailPopup } from '@src/components/popup/eventDetailPopup';
 import { EventFormPopup } from '@src/components/popup/eventFormPopup';
 import { SeeMoreEventsPopup } from '@src/components/popup/seeMoreEventsPopup';
 import { useDispatch } from '@src/contexts/calendarStore';
-import { LayoutContainerProvider } from '@src/contexts/layoutContainerRef';
+import { LayoutContainerProvider } from '@src/contexts/layoutContainer';
 import { cls, toPercent } from '@src/helpers/css';
 import { useDOMNode } from '@src/hooks/common/domNode';
 import { noop } from '@src/utils/noop';
